test: add new brand token example test, rename (reorder) examples, tidy up
