+++ components/core/examples/50-tokens-type.js
