+++ components/core/examples/80-mergeWith.js
